Deriving stdtype set into a Nodeset to help keep track of nodes
It got complicated. Node class has two list of Nodes for enter and
leave. Graph has a dictionary of Nodes for keeping track of all the
Nodes in the graph. List doesn't offer built in dedup, which is super
helpful when minimizing the graph. Sets would be perfect, but it doesn't
offer an accessor. A graph does, but it has advantages and disadvantages
as well. In the end, I am deriving a set in a new class, Nodeset.
Hopefully this approach will work